Moving sample distribution to not be async
