+++ _infra/helm/collection-exercise/values.yaml
@@ -8,10 +8,6 @@ image:
   pullPolicy: Always
 
 crons:
-  distributeSampleUnit:
-    name: collection-exercise-sample-unit-distribution
-    cron: "* * * * *"
-    target: "cron/sample-unit-distribution"
   validateSampleUnit:
     name: collection-exercise-sample-unit-validation
     cron: "* * * * *"
